requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
alsa-lib=1.2.11=hd590300_1
aom=3.8.2=h59595ed_0
asttokens=2.4.1=pyhd8ed1ab_0
attr=2.5.1=h166bdaf_1
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
bottleneck=1.3.7=py38ha9d4c09_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.5.1=pypi_0
cairo=1.18.0=h3faef2a_0
certifi=2024.2.2=py38h06a4308_0
charset-normalizer=3.4.1=pypi_0
cloudpickle=2.2.1=py38h06a4308_0
colorama=0.4.6=py38h06a4308_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.0.5=py38hdb19cb5_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-runtime=12.1.0=0
cudatoolkit=11.5.2=hbdc67f6_13
cycler=0.11.0=pyhd3eb1b0_0
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
debugpy=1.8.1=py38h17151c0_0
decorator=5.1.1=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h6a678d5_0
fluidsynth=2.3.4=h16f8cd4_1
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.2=h14ed4e7_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.51.0=py38h5eee18b_0
freetype=2.12.1=h4a9f257_0
gettext=0.22.5=h59595ed_2
gettext-tools=0.22.5=h59595ed_2
glib=2.80.0=hf2295e7_6
glib-tools=2.80.0=hde27a5a_6
google-auth=2.38.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
graphite2=1.3.14=h295c915_1
grpcio=1.70.0=pypi_0
gst-plugins-base=1.24.1=hfa15dee_1
gstreamer=1.24.1=h98fc4e7_1
gym=0.23.0=pypi_0
gym-notices=0.0.8=pypi_0
harfbuzz=8.3.0=h3d44ed6_0
icu=73.2=h59595ed_0
idna=3.10=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib-resources=6.1.1=pyhd3eb1b0_1
importlib_metadata=7.1.0=hd8ed1ab_0
importlib_resources=6.1.1=py38h06a4308_1
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pyhd33586a_0
ipython=8.12.2=pyh41d4057_0
jack=1.9.22=h7c63dc7_2
jedi=0.19.1=pyhd8ed1ab_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py38h578d9bd_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.21.2=h659d440_0
lame=3.100=h7b6447c_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.40=h55db66e_0
lerc=4.0.0=h27087fc_0
libasprintf=0.22.5=h661eb56_2
libasprintf-devel=0.22.5=h661eb56_2
libavif16=1.0.4=hd9d6309_2
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libcap=2.69=h0f662aa_0
libcblas=3.9.0=5_h92ddd45_netlib
libclang-cpp15=15.0.7=default_h127d8a8_5
libclang13=18.1.3=default_h5d6823c_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.1.3=0
libcups=2.3.3=h4637d8d_4
libcurand=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdb=6.2.32=h6a678d5_1
libdeflate=1.20=hd590300_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.12=hdbd6064_1
libexpat=2.6.2=h59595ed_0
libffi=3.4.4=h6a678d5_1
libflac=1.4.3=h59595ed_0
libgcc-ng=13.2.0=hc881cc4_6
libgcrypt=1.10.3=hd590300_0
libgettextpo=0.22.5=h59595ed_2
libgettextpo-devel=0.22.5=h59595ed_2
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.80.0=hf2295e7_6
libgomp=13.2.0=hc881cc4_6
libgpg-error=1.49=h4f305b6_0
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=5_h92ddd45_netlib
libllvm15=15.0.7=hb3ce162_4
libllvm18=18.1.4=h2448989_0
libmad=0.15.1b=h0b41bf4_1001
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libogg=1.3.5=h27cfd23_1
libopenblas=0.3.21=h043d6bf_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.43=h2797004_0
libpq=16.2=h33b98f1_1
libsndfile=1.2.2=hc60ed4a_1
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.45.3=h2797004_0
libstdcxx-ng=13.2.0=h95c4c6d_6
libsystemd0=255=h3516f8a_1
libtiff=4.6.0=h1dd3fc0_3
libuuid=2.38.1=h0b41bf4_0
libuv=1.44.2=h5eee18b_0
libvorbis=1.3.7=h7b6447c_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h662e7e4_0
libxml2=2.12.6=h232c23b_2
libzlib=1.2.13=hd590300_5
lz4-c=1.9.4=h6a678d5_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.3=py38h578d9bd_0
matplotlib-base=3.7.3=py38h58ed7fa_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpg123=1.32.6=h59595ed_0
mysql-common=8.3.0=hf1915f5_4
mysql-libs=8.3.0=hca2cd23_4
ncurses=6.4.20240210=h59595ed_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nspr=4.35=h6a678d5_0
nss=3.98=h1d7d5a4_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=pypi_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.3.0=hd590300_0
opusfile=0.12=h3358134_2
packaging=24.0=pyhd8ed1ab_0
pandas=2.0.3=py38h1128e8f_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.43=hcad00b1_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py38h9e66945_0
pip=24.0=pyhd8ed1ab_0
pixman=0.43.2=h59595ed_0
platformdirs=4.2.1=pyhd8ed1ab_0
ply=3.11=py38_0
portaudio=19.6.0=h7b6447c_4
portmidi=2.0.4=h7c63dc7_2
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=5.29.3=pypi_0
psutil=5.9.8=py38h01eb140_0
ptyprocess=0.7.0=pyhd3deb0d_0
pulseaudio-client=17.0=hb77b528_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pygame=2.5.2=py38h42b80ce_2
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.10=py38h6a678d5_0
pyqt5-sip=12.13.0=py38h5eee18b_0
python=3.8.19=hd12c33a_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=4_cp38
pytorch=1.11.0=py3.8_cuda11.5_cudnn8.3.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py38h06a4308_0
pyzmq=26.0.2=py38ha44f8e3_0
qt-main=5.15.8=hc9dc06e_21
rav1e=0.6.6=he8a937b_2
readline=8.2=h8228510_1
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
scipy=1.8.0=py38h56a6a73_1
sdl2=2.30.2=hdbcbe63_0
sdl2_image=2.8.0=h53e0c72_0
sdl2_mixer=2.6.3=h8830914_1
sdl2_ttf=2.22.0=ha14f88b_1
setuptools=69.5.1=pyhd8ed1ab_0
sip=6.7.12=py38h6a678d5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h2c6b66d_0
stack_data=0.6.2=pyhd8ed1ab_0
svt-av1=2.0.0=h59595ed_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.0.1=py38h06a4308_0
tornado=6.4=py38h01eb140_0
tqdm=4.63.0=pyhd3eb1b0_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.11.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.6=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xcb-util=0.4.0=hd590300_1
xcb-util-image=0.4.0=h8ee46fc_1
xcb-util-keysyms=0.4.0=h8ee46fc_1
xcb-util-renderutil=0.3.9=hd590300_1
xcb-util-wm=0.4.1=h8ee46fc_1
xkeyboard-config=2.41=hd590300_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.9=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xf86vidmodeproto=2.3.1=h7f98852_1002
xorg-xproto=7.0.31=h27cfd23_1007
xz=5.4.6=h5eee18b_0
zeromq=4.3.5=h59595ed_2
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hc292b87_0
